market order service, changed longs to ints for object ids

+++ src/main/java/com/natewilliford/fixer/api/ResourceTransferServerResource.java
@@ -13,15 +13,14 @@ public class ResourceTransferServerResource extends ServerResource {
         User user = Main.game.users.getForName(getChallengeResponse().getIdentifier());
 
         try {
-            long objectId = Long.parseLong(getAttribute("objectId"));
+            int objectId = Integer.parseInt(getAttribute("objectId"));
             int type = Integer.parseInt(getAttribute("type"));
-            long toObjectId = Long.parseLong(getAttribute("toObjectId"));
+            int toObjectId = Integer.parseInt(getAttribute("toObjectId"));
             long amount = Long.parseLong(getAttribute("amount"));
 
             Main.game.transferResource(user.getId(), objectId, toObjectId, type, amount);
         } catch (Exception e) {
             getResponse().setStatus(Status.CLIENT_ERROR_BAD_REQUEST);
-//            return null;
         }
     }
 }
